Mark file as processed and a bit of refactor - now yes
