+++ matcher/data/models/sound_recording.py
@@ -1,8 +1,6 @@
 from sqlalchemy import Column, Integer, Text, text
-from sqlalchemy.ext.declarative import declarative_base
 
-Base = declarative_base()
-metadata = Base.metadata
+from .base import Base
 
 
 class SoundRecording(Base):
